requirements.txt
beautifulsoup4
duckduckgo-search
langchain==0.0.354
langchain-community
python-dotenv
urllib3==1.26.18